move to gradle; change some files
