--- ArcaneAPI/pom.xml
@@ -1,41 +1,0 @@
-<project xmlns="http://maven.apache.org/POM/4.0.0" xmlns:xsi="http://www.w3.org/2001/XMLSchema-instance" xsi:schemaLocation="http://maven.apache.org/POM/4.0.0 http://maven.apache.org/xsd/maven-4.0.0.xsd">
-  <modelVersion>4.0.0</modelVersion>
-  <parent>
-    <groupId>com.arcaneminecraft</groupId>
-    <artifactId>ArcanePlugins</artifactId>
-    <version>2.1</version>
-  </parent>
-  <groupId>com.arcaneminecraft.api</groupId>
-  <artifactId>ArcaneAPI</artifactId>
-  <version>2.1.0-SNAPSHOT</version>
-  <dependencies>
-    <dependency>
-      <groupId>com.sparkjava</groupId>
-      <artifactId>spark-core</artifactId>
-      <version>2.6.0</version>
-    </dependency>
-  </dependencies>
-  <build>
-    <plugins>
-      <plugin>
-        <groupId>org.apache.maven.plugins</groupId>
-        <artifactId>maven-shade-plugin</artifactId>
-        <version>2.4.3</version>
-        <executions>
-          <execution>
-            <phase>package</phase>
-            <goals>
-              <goal>shade</goal>
-            </goals>
-            <configuration>
-              <relocation>
-                <pattern>net.dv8tion.jda</pattern>
-                <shadedPattern>com.arcaneminecraft.dependencies.jda</shadedPattern>
-              </relocation>
-            </configuration>
-          </execution>
-        </executions>
-      </plugin>
-    </plugins>
-  </build>
-</project>
