--- ArcaneAPI/src/main/java/com/arcaneminecraft/api/ColorPalette.java
@@ -1,18 +1,0 @@
-package com.arcaneminecraft.api;
-
-import net.md_5.bungee.api.ChatColor;
-
-public class ColorPalette {
-	public static final ChatColor HEADING = ChatColor.GOLD;
-	public static final ChatColor CONTENT = ChatColor.GRAY;
-	public static final ChatColor FOCUS = ChatColor.WHITE;
-	public static final ChatColor NEGATIVE = ChatColor.RED;
-	public static final ChatColor POSITIVE = ChatColor.GREEN;
-	public static final ChatColor MAP = ChatColor.DARK_GREEN;
-	public static final ChatColor DONOR = ChatColor.DARK_AQUA;
-	public static final ChatColor CHAT_MOD = ChatColor.DARK_GREEN;
-	public static final ChatColor MOD = ChatColor.BLUE;
-	public static final ChatColor ADMIN = ChatColor.RED;
-	public static final ChatColor RESET = ChatColor.RESET;
-	public static final ChatColor META = ChatColor.YELLOW;
-}
