+++ src/main/java/com/arcaneminecraft/api/ColorPalette.java
@@ -1,0 +1,19 @@
+package com.arcaneminecraft.api;
+
+import net.md_5.bungee.api.ChatColor;
+
+public class ColorPalette {
+    // Not set to final on purpose
+    public static ChatColor HEADING = ChatColor.GOLD;
+    public static ChatColor CONTENT = ChatColor.GRAY;
+    public static ChatColor FOCUS = ChatColor.WHITE;
+    public static ChatColor NEGATIVE = ChatColor.RED;
+    public static ChatColor POSITIVE = ChatColor.GREEN;
+    public static ChatColor MAP = ChatColor.DARK_GREEN;
+    public static ChatColor DONOR = ChatColor.DARK_AQUA;
+    public static ChatColor CHAT_MOD = ChatColor.DARK_GREEN;
+    public static ChatColor MOD = ChatColor.BLUE;
+    public static ChatColor ADMIN = ChatColor.RED;
+    public static ChatColor RESET = ChatColor.RESET;
+    public static ChatColor META = ChatColor.YELLOW;
+}
