+++ src/main/java/com/arcaneminecraft/api/SpigotArcAPI.java
@@ -1,0 +1,17 @@
+package com.arcaneminecraft.api;
+
+import org.bukkit.entity.Player;
+import org.bukkit.plugin.java.JavaPlugin;
+import org.bukkit.plugin.messaging.PluginMessageListener;
+
+public class SpigotArcAPI extends JavaPlugin implements PluginMessageListener {
+    @Override
+    public void onEnable() {
+
+    }
+
+    @Override
+    public void onPluginMessageReceived(String channel, Player player, byte[] message) {
+
+    }
+}
