+++ src/main/resources/bungee.yml
@@ -1,0 +1,5 @@
+name: ArcaneAPI
+main: com.arcaneminecraft.api.BungeeArcAPI
+version: ${project.version}
+description: Arcane's Color, Web, and Utility API
+author: SimonOrJ
