+++ src/main/resources/plugin.yml
@@ -1,0 +1,8 @@
+name: ArcaneAPI
+main: com.arcaneminecraft.api.ArcaneAPI
+version: ${project.version}
+description: Arcane's Color, Web, and Utility API
+authors: [Agentred100, jugglingman456, Morios, SimonOrJ]
+website: https://arcaneminecraft.com/
+
+load: STARTUP
